# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Mini Project tentang penyewaan alat-alat bertema lingkungan mulai dari alat perawatan tanaman, kebersihan lingkungan, alat hiking dan camping, dan alat proses daur ulang.",
        "title": "Chandra Mini Project",
        "contact": {
            "url": "http://www.swagger.io/support"
        },
        "version": "1.0"
    },
    "basePath": "/v2",
    "paths": {
        "/admin/confirm/{id}": {
            "put": {
                "description": "Confirmation Data Rental by Admin",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Confirmation Rental"
                ],
                "summary": "Confirmation Data Rental",
                "operationId": "Confirmiation-Rental",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.RentConfirmRespond"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/admin/info": {
            "get": {
                "description": "Admin can see all info user that have Returned and not returned",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Info data of the user who rented"
                ],
                "summary": "Get All Info Returned and not returned",
                "operationId": "Info-Rerutned",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.RentalInfoRespond"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/admin/info{id}": {
            "put": {
                "description": "Admin can confirm that user have returned equipment and not returned it",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Info data of the user who rented"
                ],
                "summary": "Update Data Returned and not returned equipment",
                "operationId": "Confirmation-Rerutned-data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.RentalInfoRespond"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/confirm": {
            "get": {
                "description": "Get All Data Confirmiation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Confirmation Rental"
                ],
                "summary": "Get All Data Confirmiation",
                "operationId": "Get-All-Confirmiation",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.RentConfirmRespond"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "post": {
                "description": "Post Data Confirmiation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Confirmation Rental"
                ],
                "summary": "Post Data Confirmiation",
                "operationId": "Post-Confirmiation",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.RentConfirmRespond"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/confirm/user": {
            "get": {
                "description": "User Get All Data Confirmiation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Confirmation Rental"
                ],
                "summary": "Get All Data Confirmation for User",
                "operationId": "Get-All-Confirmiation-For-User",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.RentConfirmRespond"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/confirm/user/{id}": {
            "delete": {
                "description": "User Can cancel confirmation before it got reviewed by admin",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Confirmation Rental"
                ],
                "summary": "Cancel Data Confirmation for User",
                "operationId": "Cancel-Data-Confirmation-For-User",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.RentConfirmRespond"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/login": {
            "post": {
                "description": "User can login with email and password",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Register"
                ],
                "summary": "Login for user",
                "operationId": "Login-User",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "User can register with name, email and password",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Register"
                ],
                "summary": "Signup for user",
                "operationId": "Register-User",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "prototype_api_controller_rent_confirm_response.UserData": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "response.LoginResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john@gmail.com"
                },
                "id": {
                    "type": "string",
                    "example": "uuid"
                },
                "token": {
                    "type": "string",
                    "example": "token"
                }
            }
        },
        "response.RentConfirmRespond": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "Jl. Setiabudi No. 1, Jakarta, Indonesia"
                },
                "admin_id": {
                    "type": "string",
                    "example": "uuid"
                },
                "date_start": {
                    "type": "string",
                    "example": "2024-00-00 00:00:00"
                },
                "delivery": {
                    "type": "boolean",
                    "example": true
                },
                "duration": {
                    "type": "integer",
                    "example": 1
                },
                "fee": {
                    "type": "integer",
                    "example": 100000
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "payment_method": {
                    "type": "string",
                    "example": "http://cloudinary.com/photo/2016/03/31/15/32/robot-1295393_960_720.png"
                },
                "rent": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.RentDetails"
                    }
                },
                "return_time": {
                    "type": "string",
                    "example": "2024-00-00 00:00:00"
                },
                "status": {
                    "type": "string",
                    "example": "pending"
                },
                "user_data": {
                    "$ref": "#/definitions/prototype_api_controller_rent_confirm_response.UserData"
                }
            }
        },
        "response.RentDetails": {
            "type": "object",
            "properties": {
                "equipment_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "response.RentDetailsInfo": {
            "type": "object",
            "properties": {
                "equipment_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "response.RentalInfoRespond": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "Jl. Setiabudi No. 1, Jakarta, Indonesia"
                },
                "duration": {
                    "type": "integer",
                    "example": 1
                },
                "fee": {
                    "type": "integer",
                    "example": 100000
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "rent": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.RentDetailsInfo"
                    }
                },
                "return_time": {
                    "type": "string",
                    "example": "2024-11-10 09:00:00"
                },
                "status": {
                    "type": "string",
                    "example": "Returned"
                },
                "user_data": {
                    "$ref": "#/definitions/prototype_api_controller_rent_confirm_response.UserData"
                }
            }
        },
        "response.UserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john@gmail.com"
                },
                "id": {
                    "type": "string",
                    "example": "uuid"
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "username": {
                    "type": "string",
                    "example": "johndoe"
                }
            }
        }
    }
}